fix stats functions to work with fold change fun
Added formulas to the output list of three stats functions so that the mt_post_fold_change function can be called on them. Added an error message in the mt_post_fold_change function if a formula is not found in the output list of the results of the provided stat_name.