bug fix: mt_modify_agg_pathways
Fixed a bug in mt_modify_agg_pathways that resulted in the newly created pathways SE object having assay rows and rowData rows of different length. The issue arose for the case of nested pathways. Solution was to check that list of unique (and unnested) pathways is equal to length of unique pathway_col entries. If not, just use pathway ID and names, do not attempt to copy over any other columns from rowData.